modify routes pt1
