+++ backend/routes/api/booking.js
@@ -1,4 +1,4 @@
-const router = require("express").Router();
+const router = require("express").Router({ mergeParams: true });
 const { requireAuth } = require("../../utils/auth");
 const { Booking, Spot } = require("../../db/models");
 const { check } = require("express-validator");
@@ -35,7 +35,7 @@ const validateBooking = [
 ========================================
 */
 
-router.get("/", requireAuth, async (req, res, next) => {
+router.get("/current", requireAuth, async (req, res, next) => {
   const uid = req.user.id;
 
   try {
@@ -55,6 +55,110 @@ router.get("/", requireAuth, async (req, res, next) => {
   }
 });
 
+// get all bookings for a spot based on spot id
+router.get("/:spotId/bookings", async (req, res, next) => {
+  const spotId = Number(req.params.spotId);
+  const uid = req.user.id;
+
+  try {
+    const spot = await Spot.findByPk(spotId);
+    if (!spot) {
+      const err = new Error("Spot couldn't be found");
+      err.status = 404;
+      return next(err);
+    }
+    // is the user the owner of the spot?
+    const isOwner = spot.ownerId === uid;
+
+    let bookings;
+
+    // if the user is the owner of the spot, include all details
+    if (isOwner) {
+      bookings = await Booking.findAll({
+        where: { spotId },
+        include: [
+          {
+            model: User,
+            attributes: userAttributes, // only has id, firstName, lastName
+            as: "User",
+          },
+        ],
+      });
+
+      return res.json({ Bookings: bookings });
+    }
+
+    // if the user is not the owner of the spot, only include basic details
+    bookings = await Booking.findAll({
+      where: { spotId },
+      attributes: ["spotId", "startDate", "endDate"],
+    });
+    return res.json({ Bookings: bookings });
+  } catch (error) {
+    next(error);
+  }
+});
+
+// create a booking from a spot based on spot id
+router.post("/", requireAuth, async (req, res) => {
+  const ownerId = req.user.id;
+  const { startDate, endDate } = req.body;
+  const spotId = Number(req.params.spotId);
+  try {
+    // check if the spot exists
+    const spot = await Spot.findByPk(spotId);
+    if (!spot) {
+      const err = new Error("Spot Image couldn't be found");
+      err.status = 404;
+      next(err);
+    }
+
+    // check if the user is the owner of the spot
+    if (spot.ownerId === ownerId) {
+      const err = new Error("Spot must not belong to user");
+      err.status = 403;
+      next(err);
+    }
+
+    // booking conflicts with any existing booking
+    const bookingConflicts = await Booking.findAll({
+      where: {
+        spotId,
+        [Op.or]: [
+          {
+            startDate: {
+              [Op.between]: [startDate, endDate],
+            },
+          },
+          {
+            endDate: {
+              [Op.between]: [startDate, endDate],
+            },
+          },
+        ],
+      },
+    });
+
+    if (bookingConflicts.length > 0) {
+      const err = new Error(
+        "Sorry, this spot is already booked for the specified dates"
+      );
+      err.status = 403;
+      return next(err);
+    }
+
+    const booking = await Booking.create({
+      spotId,
+      userId: ownerId,
+      startDate,
+      endDate,
+    });
+    res.status(201).json(booking);
+  } catch (e) {
+    next(e);
+  }
+});
+
 /*
 ========================================
    edit a booking
@@ -64,7 +168,7 @@ router.get("/", requireAuth, async (req, res, next) => {
 router.put(
   "/:bookingId",
   requireAuth,
-  validateBooking,
+
   async (req, res, next) => {
     const userId = req.user.id;
     const bookingId = Number(req.params.bookingId);
