+++ backend/routes/api/index.js
@@ -5,6 +5,8 @@ const { restoreUser } = require("../../utils/auth.js");
 const spotsRouter = require("./spots.js");
 const bookingRouter = require("./booking.js");
 const reviewsRouter = require("./reviews.js");
+const spotImagesRouter = require("./spot-images.js");
+const reviewImagesRouter = require("./review-images.js");
 // Connect restoreUser middleware to the API router
 // If current user session is valid, set req.user to the user in the database
 // If current user session is not valid, set req.user to null
@@ -16,6 +18,8 @@ router.use("/spots", spotsRouter);
 router.use("/users", usersRouter);
 router.use("/booking", bookingRouter);
 router.use("/reviews", reviewsRouter);
+router.use("/spot-images", spotImagesRouter);
+router.use("/review-images", reviewImagesRouter);
 
 router.post("/test", (req, res) => {
   res.json({ requestBody: req.body });
